libobs: Monitoring deduplication for default devices
This adds comparison to default devices to the monitoring deduplication.
When a user picks a default device, the device_id setting is 'default',
which prevents any comparison.
We fix by adding to the three audio output capture sources a new
'default_device_id' setting which stores the real device_id of the
default device, which enables comparisons.
For macOS, some special care is taken because the devices list differ
for 'Desktop Audio' and 'monitoring' since coreaudio sdk has no pure
audio capture; so 'default' in the two lists do not match in general.
The coreaudio monitoring devices_match function is then used to compare
'device_id' (or 'default_device_id' if device_id is 'default') with the
monitoring id.

Signed-off-by: pkv <pkv@obsproject.com>