libobs/audio_monitoring: Allow monitoring of muted sources
This allows the monitoring of muted sources.
In previous monitoring_type API, it was possible to monitor a visible
source without outputting its audio.
With the new API, visibility is now tied directly to what is output.
There are therefore two options now:
- either hide the source and monitor it (by displaying inactive sources
in the mixer);
- or mute the source, show it and monitor it.
The first option is already possible; we allow now the second option.
The principle used is that we should be able to monitor any vue-meter
which shows activity.
Since a muted source as well as a hidden one may show audio activity,
we must allow their monitoring.

Signed-off-by: pkv <pkv@obsproject.com>