libobs: New monitoring_type API
The monitoring_type API has been creating headaches because it couples
the monitoring with the output of audio (to streams, recordings) ...
There are three obs_monitoring_type states:
- none,
- monitor_only,
- monitor_and_output.
This leads to difficulties on both UI side and also backend and can
be a brake to a better UX. A user might see for instance active meters
but be unable to monitor them because the source is muted (to the
outputs) or inactive. But the user might want to be able to prepare
their audio source and to monitor them before having them active or
unmuted. This kind of use is rendered very complex on the backend and UI
because of the coupling of monitoring and output.

The purpose of this commit is to decouple monitoring and output by
removing the OBS_MONITORING_TYPE_MONITOR_ONLY; the enum now has only:
- OBS_MONITORING_TYPE_NONE == 0,
- OBS_MONITORING_TYPE_MONITOR == 2.

The migration is done in this manner:
- OBS_MONITORING_TYPE_MONITOR_ONLY == 1 is migrated to
- OBS_MONITORING_TYPE_MONITOR + visibility == false.
This implies that the corresponding source will not output its audio.
In order to see the meter for the source, one will have to show the
inactive sources in the mixer.

The motivation for this API change is to allow for more natural uses of
the headphone button (so better UX).
Indeed the meters still display activity even if a source is:
- muted,
- or inactive (one has clicked the 'show' button to the 'hide' state).
(In the latter case, the mixer must be set to display inactive sources.)

But since the meters may show activity, the user might want to monitor
their audio.
The decoupling of monitoring state and output allows to monitor an
inactive source, helping the user to prepare a visibility transition to
it.
Or, the source might be active but user_muted. The latter ensures audio
does not go to outputs. This also allows the user to prepare an unmute
transition by letting them to monitor the said source (provided
monitoring is not disabled by a muted state, see next commit which fixes
that).

Signed-off-by: pkv <pkv@obsproject.com>