UI: On Windows fix monitoring device id update
After Windows updates, the audio devices may have their GUID changed.
This breaks monitoring initialization in OBSBasic::OBSInit().
We fix that by checking the devices names; if there is a match, we
update to the new GUID. The check allows for partial matches to cover
the case of device names being appended a number by WIndows.
If several matches are detected, nothing is done.